Major rework
For readability, ease-of-use & modularity for enterprise applications. Secondarily, the previous version ran into a lot of errors when attempts were made to deploy on DO-droplets for k8s-based enterprise cloud-native back-ends. This new variation encompasses recent changes in the landscape, and as well makes it easier for devs to make direct changes (hence, modularity) to fit their platform/projects specific use-cases.